superai-platform/pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "superai-platform"
version = "1.0.0"
description = "超級AI自開發雲原生平台 - 量子-AI混合框架與完整科學計算生態"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Enterprise Engineering Team"},
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
    # === Core Web Framework ===
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "gunicorn>=21.2.0",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "websockets>=12.0",

    # === Database & ORM ===
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",

    # === Authentication & Security ===
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=42.0.0",

    # === Caching & Message Queue ===
    "redis[hiredis]>=5.0.1",
    "celery[redis]>=5.3.6",
    "kombu>=5.3.4",

    # === Configuration & Validation ===
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",

    # === Logging & Monitoring ===
    "structlog>=24.1.0",
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",

    # === Quantum Computing ===
    "qiskit>=1.0.0",
    "qiskit-aer>=0.13.0",
    "qiskit-algorithms>=0.3.0",
    "qiskit-ibm-runtime>=0.17.0",

    # === AI & Machine Learning ===
    "tensorflow>=2.15.0",
    "tensorflow-quantum>=0.7.3; python_version<'3.12'",
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "openai>=1.6.0",

    # === Vector Database ===
    "chromadb>=0.4.22",
    "sentence-transformers>=2.3.1",
    "pinecone-client>=3.0.0",
    "faiss-cpu>=1.7.4",

    # === Scientific Computing ===
    "numpy>=1.26.3",
    "pandas>=2.1.5",
    "scipy>=1.12.0",
    "scikit-learn>=1.4.0",
    "matplotlib>=3.8.2",
    "seaborn>=0.13.1",
    "sympy>=1.12",
    "statsmodels>=0.14.1",

    # === Search Engine ===
    "elasticsearch[async]>=8.12.0",

    # === Utilities ===
    "pyyaml>=6.0.1",
    "orjson>=3.9.12",
    "tenacity>=8.2.3",
    "aiofiles>=23.2.1",
    "jinja2>=3.1.3",
    "click>=8.1.7",
    "rich>=13.7.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "httpx>=0.26.0",
    "factory-boy>=3.3.0",
    "faker>=22.0.0",
    "black>=24.1.0",
    "isort>=5.13.2",
    "mypy>=1.8.0",
    "ruff>=0.1.14",
    "pre-commit>=3.6.0",
    "bandit>=1.7.7",
    "safety>=2.3.5",
]
docs = [
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.5.3",
    "mkdocstrings[python]>=0.24.0",
]

[project.scripts]
superai = "src.presentation.api.main:cli"
superai-worker = "src.infrastructure.tasks.worker:main"
superai-health = "tools.health_check:main"
yaml2json = "tools.yaml2json:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "quantum: Quantum computing tests",
    "slow: Slow running tests",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "DTZ", "T20", "ICN"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/migrations/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true